Thicken cantilever height to match POLY555
Hopefully this will cut down on noise. Caveat is that the enclosure now
needs to be sliggghhhtly taller, otherwise the nut juts into enclosure
top.

New min_screw_top_clearance arg ensures there's room for the screw to
clear the nut, and its seemingly arbitrary value is to keep the derived
cantilever_length at 4.6, which is what it was at before.

Also added a TODO to try the cantilever_length from previous models,
which are stored in constants for now.