Move battery_holder wire_channel to bottom
Previous front channel was too big to print well on wall w/o support or
changing the wall depth. It was also too tall, but that's less of an
issue.

Some commercial battery holders put this channel on/near the bottom,
running the cable in the gap between the batteries.

Worth a try!