Fix impatient burn_chip pause
After a couple runs it kinda has a mind of its own, maybe from
capturing keypresses switching windows? Unclear.

Regardless, the burn process is destructive enough that waiting for ANY
key press is probably too generous. Waiting instead of a specific key
is narrow enough and seems to be performing as expected.