Make led_exposure less obvious
On light plastics, a shade_depth of 3x DEFAULT_DFM_LAYER_HEIGHT has a
clearly visible inner shadow.

In testing with yellow PLA, it's around 5x that the outline is no longer
defined enough to see. At 7x it's completely invisible.

I'm worried that 5x is too much for darker plastics, which will be more
common in my prints, so compromising at 4x.